Moves all trace reader classes to org.cloudsimplus.traces package including TraceReader, TraceReaderAbstract and SwfWorkloadFileReader
