+++ src/main/java/org/cloudsimplus/traces/google/GoogleMachineEventsTraceReader.java
@@ -32,9 +32,9 @@ import org.cloudsimplus.core.CloudSimTag;
 import org.cloudsimplus.datacenters.Datacenter;
 import org.cloudsimplus.hosts.Host;
 import org.cloudsimplus.resources.Pe;
+import org.cloudsimplus.traces.TraceReaderAbstract;
 import org.cloudsimplus.util.BytesConversion;
 import org.cloudsimplus.util.ResourceLoader;
-import org.cloudsimplus.util.TraceReaderAbstract;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
