+++ src/main/java/org/cloudsimplus/traces/google/GoogleTaskUsageTraceReader.java
@@ -30,9 +30,9 @@ import org.cloudsimplus.core.CloudSimTag;
 import org.cloudsimplus.core.Simulation;
 import org.cloudsimplus.core.events.CloudSimEvent;
 import org.cloudsimplus.listeners.EventInfo;
+import org.cloudsimplus.traces.TraceReaderAbstract;
 import org.cloudsimplus.util.ResourceLoader;
 import org.cloudsimplus.util.TimeUtil;
-import org.cloudsimplus.util.TraceReaderAbstract;
 import org.cloudsimplus.utilizationmodels.UtilizationModel;
 import org.cloudsimplus.utilizationmodels.UtilizationModelDynamic;
 import org.cloudsimplus.utilizationmodels.UtilizationModelFull;
