+++ src/test/java/org/cloudsimplus/util/SwfWorkloadFileReaderTest.java
@@ -1,6 +1,7 @@
 package org.cloudsimplus.util;
 
 import org.cloudsimplus.cloudlets.Cloudlet;
+import org.cloudsimplus.traces.SwfWorkloadFileReader;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
